move more functional throttle code to the throttle.cc file.  Not quite a class yet.  But cleaner.
