+++ util/env_posix.cc
@@ -675,47 +675,6 @@ class PosixEnv : public Env {
 
   virtual int GetBackgroundBacklog() const {return(bg_backlog_);};
 
-  virtual void SetWriteRate(uint64_t Micros, uint64_t Keys, bool IsLevel0)
-  {
-      int backlog;
-
-      backlog=GetBackgroundBacklog();
-
-      if (IsLevel0)
-      {
-          pthread_mutex_lock(&gThrottleMutex);
-          gThrottleData[0].m_Micros+=Micros;
-          gThrottleData[0].m_Keys+=Keys;
-          gThrottleData[0].m_Backlog+=backlog;
-          gThrottleData[0].m_Compactions+=1;
-          pthread_mutex_unlock(&gThrottleMutex);
-
-          gPerfCounters->Add(ePerfThrottleMicros0, Micros);
-          gPerfCounters->Add(ePerfThrottleKeys0, Keys);
-          gPerfCounters->Add(ePerfThrottleBacklog0, backlog);
-          gPerfCounters->Inc(ePerfThrottleCompacts0);
-      }   // if
-
-      else
-      {
-          pthread_mutex_lock(&gThrottleMutex);
-          gThrottleData[1].m_Micros+=Micros;
-          gThrottleData[1].m_Keys+=Keys;
-          gThrottleData[1].m_Backlog+=backlog;
-          gThrottleData[1].m_Compactions+=1;
-          pthread_mutex_unlock(&gThrottleMutex);
-
-          gPerfCounters->Add(ePerfThrottleMicros1, Micros);
-          gPerfCounters->Add(ePerfThrottleKeys1, Keys);
-          gPerfCounters->Add(ePerfThrottleBacklog1, backlog);
-          gPerfCounters->Inc(ePerfThrottleCompacts1);
-      }   // else
-
-
-      return;
-  };
-
-  virtual uint64_t GetWriteRate() const {return(gThrottleRate);};
 
  private:
   void PthreadCall(const char* label, int result) {
